requirements.txt
requests
protobuf
pyserial
